+++ src/components/actions/SiteCommands.tsx
@@ -21,7 +21,7 @@ export const SiteCommands = ({ site, server }: { site: ISite; server: IServer })
           title="Trigger Deploy Script"
           onAction={() => {
             showToast(Toast.Style.Animated, "Deploying...");
-            Site.deploy({ siteId: site.id, serverId: server.id, token }).catch(() =>
+            Site.deploy({ orgSlug: server.org_slug, siteId: site.id, serverId: server.id, token }).catch(() =>
               showToast(Toast.Style.Failure, "Failed to trigger deploy script")
             );
           }}